refactor(desktop): extract chat mastra pane controller hooks
