+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/hooks/useChatMastraPaneController/index.ts
@@ -1,0 +1,1 @@
+export { useChatMastraPaneController } from "./useChatMastraPaneController";
